+++ slider/iOS/Slider/SliderViewController.m
@@ -123,7 +123,7 @@
 }
 
 - (void) updateLabels {
-    self.moves.text = [NSString stringWithFormat: @"Moves: %d", self.board.moves];
+    self.moves.text = [NSString stringWithFormat: @"Moves: %ld", (long)self.board.moves];
     NSInteger solveTime;
     
     if ([self.board solved] == TRUE) {
@@ -145,7 +145,7 @@
         if (solveTime == 9999) {
             self.solved.text = @"Solved!";
         } else {
-            self.solved.text = [NSString stringWithFormat: @"Solved in %d seconds!", solveTime];
+            self.solved.text = [NSString stringWithFormat: @"Solved in %ld seconds!", (long)solveTime];
         }
         mydata = [self saveFilePath];
         
@@ -193,7 +193,7 @@
             num = [values objectAtIndex:i];
             if (time < [num intValue]) {
                 tempnum = [num intValue];
-                [values replaceObjectAtIndex:i withObject:[NSNumber numberWithInt:time]];
+                [values replaceObjectAtIndex:i withObject:[NSNumber numberWithInt:(int)time]];
                 time = tempnum;
                 if( newtime == -1) {
                     newtime = i;
@@ -210,7 +210,7 @@
             num = [values objectAtIndex:i];
             if (moves < [num intValue]) {
                 tempnum = [num intValue];
-                [values replaceObjectAtIndex:i withObject:[NSNumber numberWithInt:moves]];
+                [values replaceObjectAtIndex:i withObject:[NSNumber numberWithInt:(int)moves]];
                 moves = tempnum;
                 if( newmove == -1) {
                     newmove = i;
